Move tooliter to gaphor.diagram.diagramtoolbox
